Fix "test 5"
Test 4 puts an entry in the bounds tables.
Test 5 (tries to) pull the entry back out.

But, the way that both the shadow and hardware tables work,
an entry will only be pulled back out if the 'ptr' value
(the 3rd word in the hardware tables) matches.  But 'ptr'
in our case is a *RANDOM* variable.  It is exceedingly
rare to have a random pointer in a 28-bit address space
match.

So, to get better testing, we don't randomize the pointer
ever time.  Only some of the time.

Also, I think I put the 'bnd0' clearing in to helper5 as
an attempt to fix some things when the xsave stuff was
broken.  That was completly bogus, but it *MATCHED* what
the hardware was doing (init'ing bnd0).  It did not match
in the exceedingly rare case I explained above.