fix bounds table allocation
First of all, we were not checking mmap() for failure with the bounds
directory.  If a system had strict overcommit turned on and not much
RAM+swap and we try to allocate a big old table, it will fail.

Also, we were allocating 2*2^31 for the size instead of 2*1^30.  That
caused even more collisions with the overcommit code